Removed Context param form prepare_wrapped_call
