+++ libmamba/include/mamba/core/context.hpp
@@ -12,8 +12,8 @@
 #include <string>
 #include <vector>
 
-#include "mamba/core/command_params.hpp"
 #include "mamba/core/common_types.hpp"
+#include "mamba/core/context_params.hpp"
 #include "mamba/core/palette.hpp"
 #include "mamba/core/subdir_parameters.hpp"
 #include "mamba/core/tasksync.hpp"
@@ -108,14 +108,6 @@ namespace mamba
         {
             std::size_t download_threads{ 5 };
             int extract_threads{ 0 };
-        };
-
-        struct PrefixParams
-        {
-            fs::u8path target_prefix;
-            fs::u8path root_prefix;
-            fs::u8path conda_prefix;
-            fs::u8path relocate_prefix;
         };
 
         // Configurable
