+++ libmamba/include/mamba/core/context_params.hpp
@@ -8,6 +8,8 @@
 
 #include <string>
 
+#include "mamba/fs/filesystem.hpp"
+
 // TODO: having a file for this single structure is a bit of
 // overkill; this should be refactored when we have more structures
 // like this (i.e. parameters structures with no dependency on other
@@ -21,5 +23,13 @@ namespace mamba
         std::string current_command{ "mamba" };
         /** Is the Context used in a mamba or mamba executable (instead of a lib). */
         bool is_mamba_exe{ false };
+    };
+
+    struct PrefixParams
+    {
+        fs::u8path target_prefix;
+        fs::u8path root_prefix;
+        fs::u8path conda_prefix;
+        fs::u8path relocate_prefix;
     };
 }
