+++ libmamba/include/mamba/core/util.hpp
@@ -16,6 +16,7 @@
 #include <string_view>
 #include <vector>
 
+#include "mamba/core/context_params.hpp"
 #include "mamba/core/error_handling.hpp"
 #include "mamba/fs/filesystem.hpp"
 
@@ -373,7 +374,6 @@ namespace mamba
     };
 
     std::unique_ptr<TemporaryFile> wrap_call(
-        const Context& context,
         const fs::u8path& root_prefix,
         const fs::u8path& prefix,
         const std::vector<std::string>& arguments,  // TODO: c++20 replace by std::span
@@ -387,9 +387,9 @@ namespace mamba
     };
 
     PreparedWrappedCall prepare_wrapped_call(
-        const Context& context,
-        const fs::u8path& prefix,
-        const std::vector<std::string>& cmd
+        const PrefixParams& prefix_params,
+        const std::vector<std::string>& cmd,
+        WrappedCallOptions options
     );
 
     /// Returns `true` if the filename matches names of files which should be interpreted as YAML.
