+++ libmamba/src/core/transaction_context.cpp
@@ -238,7 +238,11 @@ namespace mamba
         const std::string cwd = target_prefix.string();
         options.working_directory = cwd.c_str();
 
-        auto [wrapped_command, script_file] = prepare_wrapped_call(ctx, target_prefix, command);
+        auto [wrapped_command, script_file] = prepare_wrapped_call(
+            context().prefix_params,
+            command,
+            WrappedCallOptions::from_context(ctx)
+        );
         m_pyc_script_file = std::move(script_file);
 
         LOG_INFO << "Running wrapped python compilation command " << util::join(" ", command);
