+++ libmamba/src/core/util.cpp
@@ -1557,9 +1557,9 @@ namespace mamba
     }
 
     PreparedWrappedCall prepare_wrapped_call(
-        const Context& context,
-        const fs::u8path& prefix,
-        const std::vector<std::string>& cmd
+        const PrefixParams& prefix_params,
+        const std::vector<std::string>& cmd,
+        WrappedCallOptions options
     )
     {
         std::vector<std::string> command_args;
@@ -1576,12 +1576,7 @@ namespace mamba
                 );
             }
 
-            script_file = wrap_call(
-                context.prefix_params.root_prefix,
-                prefix,
-                cmd,
-                WrappedCallOptions::from_context(context)
-            );
+            script_file = wrap_call(prefix_params.root_prefix, prefix_params.target_prefix, cmd, options);
 
             command_args = { comspec.value(), "/D", "/C", script_file->path().string() };
         }
@@ -1599,12 +1594,7 @@ namespace mamba
                 shell_path = "sh";
             }
 
-            script_file = wrap_call(
-                context.prefix_params.root_prefix,
-                prefix,
-                cmd,
-                WrappedCallOptions::from_context(context)
-            );
+            script_file = wrap_call(prefix_params.root_prefix, prefix_params.target_prefix, cmd, options);
             command_args.push_back(shell_path.string());
             command_args.push_back(script_file->path().string());
         }
